#16 refactor, movie type etc extracted
