+++ src/Services/Movie.ts
@@ -1,0 +1,10 @@
+interface Movie {
+    id: number;
+    name: string;
+    title: string;
+    release_date: string;
+    poster_path: string;
+  }
+  
+  export default Movie;
+
