+++ src/Services/searchMovieFetch.tsx
@@ -1,14 +1,8 @@
 import React, { useEffect } from 'react';
 import { baseUrl, apiKey } from '../Services/tmdbApiServices';
 import useFetch from '../Hooks/useFetch';
+import Movie from "../Services/Movie"
 
-interface Movie {
-    id: number;
-    name: string;
-    title: string;
-    release_date: string;
-    poster_path: string;
-}
 
 interface searchProps {
     query: string | null,
@@ -16,7 +10,7 @@ interface searchProps {
     genreId: string | null,
 }
 
-const searchMovieFetch = ({ query, page, genreId }: searchProps,setMovies: React.Dispatch<React.SetStateAction<Movie[]>>,) => {
+const searchMovieFetch = ({ query, page, genreId }: searchProps, setMovies: React.Dispatch<React.SetStateAction<Movie[]>>,) => {
 
     const searchUrl = `${baseUrl}search/movie?api_key=${apiKey}&query=${query}&page=${page}`;
     const genreUrl = `${baseUrl}discover/movie?api_key=${apiKey}&with_genres=${genreId}&page=${page}`;
@@ -37,4 +31,4 @@ const searchMovieFetch = ({ query, page, genreId }: searchProps,setMovies: React
     return (isLoading)
 }
 
-export default searchMovieFetch
+export default searchMovieFetch;
